fonts: Use fallback variant of Roboto
Inter is missing some glyphs that users have come to expect on Android,
so use Roboto as a fallback to fix excessive tofu in some text.

Because Inter's name has been replaced with "Roboto" for compatibility,
the fallback variants are named "Roboto Fallback".

Example of text with missing glyphs in Inter: ⅙⅐⅛⅑⅒¹½⅓¼⅕

Change-Id: I4affbde067bc51366a84a93b48a9c373567dca05
[ Adapt for AOSPA ]
Signed-off-by: Kazuki H <kazukih0205@gmail.com>